Create base grid with 16 * 16 divs
Edited HTML and CSS files to style the skeleton of the website.
Using a for loop that iterates 16 * 16 times in JS, the code creates
that many empty divs and appends them to the HTML <div.container>.
Added two different classes to the empty divs so in the future the grid
can be toggled on and off.