Edit script and add left and right containers
Added a createGridSquare() function to clean the code inside the for
loop and changed variables names to keep consistency, original container
became grid and divs became squares.
Added --grid-size property to the .square class to modify CSS via JS,
this will later allow the user to set different grid sizes and modify
the CSS variable.
Edited html to add side containers for the finished product, added a
button to the left container.