Edit new grid button
Added event listener to .size-btn to handle a prompt asking the user for
the number of squares per side for the new grid. Once entered, the
existing grid is removed and a new grid is generated with a newGridSize
in the same total space as before.
Styled new grid button and its container, changed grid border color.